Clean file structure
