+++ Python_dynamics_code/Old_Files/nodeParam.py
